Rename 'proxy' handler to 'cache'
