+++ config.example.json
@@ -26,7 +26,7 @@
         "read_timeout": 12312310,
         "write_timeout": 213412314,
 
-        "default_handlers": [ {"type": "via", "settings" : {"text": "baba 1.1"}}, { "type" : "proxy" }],
+        "default_handlers": [ {"type": "via", "settings" : {"text": "baba 1.1"}}, { "type" : "cache" }],
         "default_upstream_type": "simple",
         "default_cache_zone": "default",
