+++ handler/cache/handler.go
@@ -1,4 +1,4 @@
-package proxy
+package cache
 
 import (
 	"fmt"
