+++ handler/types.go
@@ -7,9 +7,9 @@ package handler
 import (
 	"github.com/ironsmile/nedomi/config"
 
+	"github.com/ironsmile/nedomi/handler/cache"
 	"github.com/ironsmile/nedomi/handler/dir"
 	"github.com/ironsmile/nedomi/handler/flv"
-	"github.com/ironsmile/nedomi/handler/proxy"
 	"github.com/ironsmile/nedomi/handler/status"
 	"github.com/ironsmile/nedomi/handler/throttle"
 	"github.com/ironsmile/nedomi/handler/via"
@@ -20,16 +20,16 @@ type newHandlerFunc func(*config.Handler, *types.Location, types.RequestHandler)
 
 var handlerTypes = map[string]newHandlerFunc{
 
+	"cache": func(cfg *config.Handler, l *types.Location, next types.RequestHandler) (types.RequestHandler, error) {
+		return cache.New(cfg, l, next)
+	},
+
 	"dir": func(cfg *config.Handler, l *types.Location, next types.RequestHandler) (types.RequestHandler, error) {
 		return dir.New(cfg, l, next)
 	},
 
 	"flv": func(cfg *config.Handler, l *types.Location, next types.RequestHandler) (types.RequestHandler, error) {
 		return flv.New(cfg, l, next)
-	},
-
-	"proxy": func(cfg *config.Handler, l *types.Location, next types.RequestHandler) (types.RequestHandler, error) {
-		return proxy.New(cfg, l, next)
 	},
 
 	"status": func(cfg *config.Handler, l *types.Location, next types.RequestHandler) (types.RequestHandler, error) {
